Add files via upload
Simple way to alternate row color of your SharePoint 2016 lists without the need of SP Designer or elevated admin credentials to create custom pages. Add this file to your Site Assets, or similar asset folder within your SP site. Once uploaded, link the URL of the asset within a Content Editor Web Part on your SharePoint page (there are multiple online tutorials explaining how to do this).

The default color in this file is a lime green. To change the color, simply open the file with an editor, and change the hex code to a color of your liking.