SP Accordion Menu HTML
Great piece of code to help you create a collapsible accordion menu within your SharePoint 2013 or 2016 sites. Useful when needing a navigation menu that collapses a massive collection of various SP sites across your company, multiple SP list forms, and other online resources into a small amount of real estate on your site page.

Also provides SharePoint's built-in coding for launching list forms in a Modal Dialogue window. Simply replace the URL with the newform URL of your List application, give it a title to display in the window, and provide sizing.

CSS can be customized to your company/SP Site's colors as well. The code uses hex format for coloring, and can be configured to display a 2 color gradient fill.

Pick it apart and customize as needed. Happy SharePointing!