Add files via upload
Here is some simple code that will help SharePoint pages refresh automatically. You can also simply download this html file and use it within a Content Editor Web Part to achieve the same result without having to mess with any code at all.
This is great if you have a SharePoint list that receives multiple submissions per day, and you want to view the submissions in real time as they're recieved.
The HTML piece is for Internet Explorer, and the JavaScript you see is for when pages are in Google Chrome.
This code has not been tested on any other browsers beyond Internet Explorer and Google Chrome at this time.