Add files via upload
The code shown can be embedded within a SharePoint page, where once loaded, it will close the window or page tab automatically. This can also be downloaded as an HTML file and simply embedded as a Content Editor Web Part for a “no-code” solution. If the command fails for any reason, the page will display the text "Thank you! If this window doesn't close on its own, please close this window." to indicate to end users that nothing further is needed beyond closing the window.
The recommended use of this would be to have the SharePoint page where the code is embedded be the Source URL on a SharePoint list form. For example, agents can click the SP list's modified NewForm URL with the code-embedded SP page listed at the Source URL, the agent then completes their submission, but then once submitted they are routed to this script, and it simply closes their window instead of directing them to the list itself.
Font styles, sizes, colors, and other attributes are listed with the "h" and "span" tags. Feel free to modify to fit your environment/branding.